refactor(predicates): extract QueuePredicateMixin from app.py
The predicate state (six instance dicts plus PredicateMetrics) and the
three gating methods (_apply_enqueue_predicate, _apply_dispatch_predicate,
_reenqueue_after_defer) plus the inspection / registration API
(list_predicates, predicate_for, register_predicate) all moved to a new
QueuePredicateMixin. Event-emission boilerplate was lifted into small
helpers on the mixin to remove the four repeated _emit_event(...) blocks.

app.py drops from 901 to 619 LOC; the mixin owns ~280 LOC of cohesive
predicate logic.
